# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hubbub team command-center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAdminUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Projects ───────────────────────────────────────────────────────────────
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects the current user is a member of
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{slug}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Add member to project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMemberInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{slug}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove member from project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ── Scopes ─────────────────────────────────────────────────────────────────
  /projects/{slug}/scopes:
    get:
      operationId: listScopes
      tags: [scopes]
      summary: List scopes for a project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scopes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scope"
    post:
      operationId: createScope
      tags: [scopes]
      summary: Create scope
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scope"
  /projects/{slug}/scopes/{scopeId}:
    patch:
      operationId: updateScope
      tags: [scopes]
      summary: Update scope
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: scopeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scope"
    delete:
      operationId: deleteScope
      tags: [scopes]
      summary: Delete scope
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: scopeId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Milestones ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, displayName, role, active]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: ["string", "null"]
        displayName:
          type: string
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        hourlyRateCents:
          type: ["integer", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, displayName]
      properties:
        email:
          type: string
        displayName:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        hourlyRateCents:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
    UserAdminUpdate:
      type: object
      properties:
        displayName:
          type: string
        role:
          type: string
          enum: [admin, member]
        hourlyRateCents:
          type: ["integer", "null"]
        active:
          type: boolean

    # ── Projects ─────────────────────────────────────────────────────────
    Project:
      type: object
      required: [id, name, slug, archived, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        archived:
          type: boolean
        memberCount:
          type: integer
        openItemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      type: object
      required: [id, name, slug, archived, members, scopes, milestones, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        archived:
          type: boolean
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/Scope"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        archived:
          type: boolean
    ProjectMember:
      type: object
      required: [id, projectId, userId, role, user]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: string
        role:
          type: string
          enum: [owner, member]
        user:
          $ref: "#/components/schemas/User"
    ProjectMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, member]

    # ── Scopes ────────────────────────────────────────────────────────────
    Scope:
      type: object
      required: [id, projectId, name, slug, status, order]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        slug:
          type: string
        sow:
          type: ["string", "null"]
        budgetCents:
          type: ["integer", "null"]
        status:
          type: string
          enum: [planned, active, on_hold, complete]
        startDate:
          type: ["string", "null"]
          format: date
        targetDate:
          type: ["string", "null"]
          format: date
        order:
          type: integer
        spentCents:
          type: ["integer", "null"]
    ScopeInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        sow:
          type: ["string", "null"]
        budgetCents:
          type: ["integer", "null"]
        status:
          type: string
          enum: [planned, active, on_hold, complete]
        startDate:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
    ScopeUpdate:
      type: object
      properties:
        name:
          type: string
        sow:
          type: ["string", "null"]
        budgetCents:
          type: ["integer", "null"]
        status:
          type: string
          enum: [planned, active, on_hold, complete]
        startDate:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
        order:
          type: integer

    # ── Milestones ────────────────────────────────────────────────────────
    Milestone:
      type: object
      required: [id, scopeId, name, status, order]
      properties:
        id:
          type: integer
        scopeId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [open, complete]
        order:
          type: integer
...